+++ tests/test_robust.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import numpy as np
+import pytest
 from scipy.stats import chi2
 
 import nustattools.stats as r
@@ -23,7 +24,24 @@ def test_derate_single_covariance():
             [np.nan, np.nan, 2.0, 3.0],
         ]
     )
-    assert np.abs(r.derate_covariance(cov, sigma=2) - 1.29) < 0.1
+    assert (
+        np.abs(r.derate_covariance(cov, sigma=2, whitening="mahalanobis") - 1.29) < 0.1
+    )
+    assert np.abs(r.derate_covariance(cov, sigma=2, whitening="cholesky") - 1.29) < 0.1
+
+    with pytest.raises(ValueError, match="Unknown whitening"):
+        r.derate_covariance(cov, whitening="unknown")
+
+    cov = np.array(
+        [
+            [2.0, 2.0, np.nan, np.nan],
+            [2.0, 2.0, np.nan, np.nan],
+            [np.nan, np.nan, 3.0, 2.0],
+            [np.nan, np.nan, 2.0, 3.0],
+        ]
+    )
+    with pytest.warns(UserWarning):
+        assert np.abs(r.derate_covariance(cov, sigma=2) - 1.41) < 0.1
 
 
 def test_derate_known_off_diag():
@@ -56,8 +74,13 @@ def test_derate_multi_covariance():
             [np.nan, np.nan, np.nan, 3.0],
         ]
     )
+    assert np.abs(r.derate_covariance([cov1, cov2, cov3], sigma=2) - 1.16) < 0.1
     cov4 = np.zeros((4, 4))
-    assert np.abs(r.derate_covariance([cov1, cov2, cov3, cov4], sigma=2) - 1.16) < 0.1
+    with pytest.raises(ValueError, match="is all zeros"):
+        r.derate_covariance([cov1, cov2, cov3, cov4], sigma=2)
+    cov3[-1, -1] = 0
+    with pytest.warns(UserWarning), pytest.raises(ValueError, match="is all zeros"):
+        r.derate_covariance([cov1, cov2, cov3], sigma=2)
 
 
 def test_derate_single_covariance_fit():